Fix INT4/INT8 type mismatch in segment existence queries
SELECT 1 returns INT4 (32-bit) in PostgreSQL, not INT8 (64-bit).
Changed Rust type from i64 to i32 in segment_effort_exists() and
is_segment_starred() functions.

+++ crates/tracks/src/database.rs
@@ -639,7 +639,7 @@ impl Database {
         segment_id: Uuid,
         activity_id: Uuid,
     ) -> Result<bool, AppError> {
-        let row: Option<(i64,)> = sqlx::query_as(
+        let row: Option<(i32,)> = sqlx::query_as(
             r#"
             SELECT 1 FROM segment_efforts
             WHERE segment_id = $1 AND activity_id = $2
@@ -820,7 +820,7 @@ impl Database {
         user_id: Uuid,
         segment_id: Uuid,
     ) -> Result<bool, AppError> {
-        let row: Option<(i64,)> = sqlx::query_as(
+        let row: Option<(i32,)> = sqlx::query_as(
             r#"
             SELECT 1 FROM segment_stars
             WHERE user_id = $1 AND segment_id = $2
